US5903 - DDK Foundation Remove CardsModule, add UI component routing.
