+++ src/app/ui-components/component-list/component-list.component.html
@@ -1,13 +1,23 @@
-<ul class="media-list">
-  <li class="media" *ngFor="let component of components">
-    <div class="media-image">
-      <img class="media-object"  [src]="component.image" alt="">
-    </div>
-    <div class="media-body">
-      <a routerLink="/cards" routerLinkActive="active" >
-        <h2 class="media-heading">{{ component.heading }}</h2>
-      </a>
-      {{ component.description }}
-    </div>
-  </li>
-</ul>
+<div class="jumbotron">
+  <div class="container">
+    <h1>UI Components</h1>
+    <p>Vestibulum id ligula porta felis euismod semper. Donec sed odio dui. Donec sed odio dui. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus.</p>
+  </div>
+</div>
+
+<div class="container">
+  <ddk-search-bar></ddk-search-bar>
+  <ul class="media-list">
+    <li class="media" *ngFor="let component of components">
+      <div class="media-image">
+        <img class="media-object"  [src]="component.image" alt="">
+      </div>
+      <div class="media-body">
+        <a routerLink="/ui/cards" routerLinkActive="active" >
+          <h2 class="media-heading">{{ component.heading }}</h2>
+        </a>
+        {{ component.description }}
+      </div>
+    </li>
+  </ul>
+</div>
