--- src/app/ui-components/ui-components.component.html
@@ -1,11 +1,0 @@
-<div class="jumbotron">
-  <div class="container">
-    <h1>UI Components</h1>
-    <p>Vestibulum id ligula porta felis euismod semper. Donec sed odio dui. Donec sed odio dui. Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo sit amet risus.</p>
-  </div>
-</div>
-
-<div class="container">
-  <ddk-search-bar></ddk-search-bar>
-  <ddk-component-list></ddk-component-list>
-</div>
