+++ src/app/ui-components/ui-components.module.ts
@@ -1,24 +1,25 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RouterModule, Routes } from '@angular/router';
+
+import { TabsModule } from 'ng2-bootstrap/ng2-bootstrap';
 
 import { UiComponentsComponent } from './ui-components.component';
+import { UiRoutingModule } from './ui-routing.module';
 import { SearchBarComponent } from './search-bar/search-bar.component';
 import { ComponentListComponent } from './component-list/component-list.component';
 
-import { CardsModule } from './cards/cards.module';
-
-const uiRoutes: Routes = [
-  {
-    path: 'ui-components',
-    component: UiComponentsComponent
-  }
-]
+import { CardsComponent } from './cards/cards.component';
+import { SidebarNavComponent } from './cards/sidebar-nav/sidebar-nav.component';
+import { BasicComponent } from './cards/basic/basic.component';
+import { VariationsComponent } from './cards/variations/variations.component';
+import { NoImageComponent } from './cards/no-image/no-image.component';
+import { VideoComponent } from './cards/video/video.component';
 
 @NgModule({
   imports: [
     CommonModule,
-    RouterModule.forChild(uiRoutes)
+    UiRoutingModule,
+    TabsModule
   ],
   exports: [
     UiComponentsComponent
@@ -26,7 +27,13 @@ const uiRoutes: Routes = [
   declarations: [
     UiComponentsComponent,
     SearchBarComponent,
-    ComponentListComponent
+    ComponentListComponent,
+    CardsComponent,
+    SidebarNavComponent,
+    BasicComponent,
+    VariationsComponent,
+    NoImageComponent,
+    VideoComponent
   ]
 })
 export class UiComponentsModule { }
